# lib/api-zod/src/generated/types/gameArmy.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GameArmy {
  /** @minimum 0 */
  scout: number;
  /** @minimum 0 */
  infantry: number;
  /** @minimum 0 */
  cavalry: number;
  /** @minimum 0 */
  archer: number;
  /** @minimum 0 */
  totalPower: number;
}