# lib/api-zod/src/generated/types/gameBattleReport.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameBattleReportBattleType } from './gameBattleReportBattleType';
import type { GameBattleReportResult } from './gameBattleReportResult';

export interface GameBattleReport {
  id: string;
  battleType: GameBattleReportBattleType;
  opponent: string;
  result: GameBattleReportResult;
  powerDelta: number;
  reward: string;
  createdAt: Date;
}