# lib/api-zod/src/generated/types/gameCommandInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameCommandInputAction } from './gameCommandInputAction';

export interface GameCommandInput {
  /** @minLength 1 */
  initData: string;
  action: GameCommandInputAction;
  targetId?: string;
  /** @minimum 1 */
  quantity?: number;
  /**
     * @minLength 3
     * @maxLength 32
     */
  name?: string;
  /**
     * @minLength 2
     * @maxLength 6
     */
  tag?: string;
}