# lib/api-zod/src/generated/types/gameResearch.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameResource } from './gameResource';

export interface GameResearch {
  id: string;
  name: string;
  /** @minimum 0 */
  level: number;
  description: string;
  researchCost: GameResource;
}