# lib/api-zod/src/generated/types/gameResource.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GameResource {
  /** @minimum 0 */
  food: number;
  /** @minimum 0 */
  wood: number;
  /** @minimum 0 */
  stone: number;
  /** @minimum 0 */
  iron: number;
}