# lib/api-zod/src/generated/types/gameWorldEvent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameWorldEventStatus } from './gameWorldEventStatus';

export interface GameWorldEvent {
  id: string;
  title: string;
  description: string;
  /** @minimum 0 */
  progress: number;
  /** @minimum 1 */
  goal: number;
  endsAt: Date;
  reward: string;
  status: GameWorldEventStatus;
}