# lib/api-zod/src/generated/types/gameWorldMap.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameMapTile } from './gameMapTile';

export interface GameWorldMap {
  centerX: number;
  centerY: number;
  tiles: GameMapTile[];
}